requirements.txt
!pip install --upgrade pip
!pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
!pip install transformers==4.27.0
!pip install sentence-transformers==2.2.2
!pip install huggingface-hub==0.19.4
!pip install accelerate==0.20.3
!pip install langchain langchain-community faiss-cpu python-dotenv ipywidgets
!pip install pyngrok
!pip install flask
!pip install flask_cors
!pip install kaggle
!pip install accelerate

langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.36.0
faiss-cpu>=1.7.4
ipywidgets>=8.1.1
pandas>=2.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2